Add attribute observer & generic attributes model
